feature/add-optional-kwargs: Added optional kwargs
Most of the init arguments in the ``HTML2Text`` class are hardcoded in
constants and modifiable only by the cli, not through the library
usage. This adds the possibility to pass kwargs through the function
call ``html2text`` or class init.

Please note that the commit contains syntax that is not recognizable
by ``mypy``, but is correct. Note: https://github.com/python/mypy/issues/5719