Expand use of dict get method to simplify code
In the code, there was often both a check for whether a certain key was in a dict and a conditional access of the corresponding value. This can be elegantly combined in a single call to the dict get method, whose default value if the key is not in the dict is None.
Using the get method instead of the check and conditional access makes it possible to simplify the code in various places.
This is done so in this patch.